@@ -1,11 +1,10 @@
 name: picker_view
 description: A Flutter picker view like iOS UIPickerView support multiple pickers.
-version: 1.0.1
+version: 1.0.2
 homepage: https://github.com/sososdk/picker_view.git
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=3.5.4 <4.0.0"
 
 dependencies:
   flutter:
